+++ src/snippen_sms/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-__version__ = "0.16.2"
+__version__ = "0.17.0"
 
 from snippen_sms.client import (
     SnippenApiError,